исправляет замечания от наставника
