+++ data.php
@@ -1,0 +1,10 @@
+<?php
+$projects = ['Все', 'Входящие', 'Учеба', 'Работа', 'Домашние дела', 'Авто'];
+$tasks = [
+['task' => 'Собеседование в IT компании', 'date' => '01.06.2018', 'category' => 'Работа', 'status' => 'Нет'],
+['task' => 'Выполнить тестовое задание', 'date' => '25.05.2018', 'category' => 'Работа', 'status' => 'Нет'],
+['task' => 'Сделать задание первого раздела', 'date' => '21.04.2018', 'category' => 'Учеба', 'status' => 'Да'],
+['task' => 'Встреча с другом', 'date' => '22.04.2018', 'category' => 'Входящие', 'status' => 'Нет'],
+['task' => 'Купить корм для кота', 'date' => 'Нет', 'category' => 'Домашние дела', 'status' => 'Нет'],
+['task' => 'Заказать пиццу', 'date' => 'Нет', 'category' => 'Домашние дела', 'status' => 'Нет'],
+];
